Database changes and environment implementation
