+++ package.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "concurrently": "^7.4.0",
+    "dotenv": "^16.0.3",
     "express": "^4.18.1",
     "http-proxy-middleware": "^2.0.6",
     "mongoose": "^6.6.5",
